Moved ITest to the test api.
